requirements.txt
streamlit
networkx
pandas
numpy
plotly
pyvis
scipy
python-louvain
streamlit-authenticator
pyyaml
